+++ api/multer.ts
@@ -1,0 +1,40 @@
+import multer from "multer";
+import { logger } from "../shared/logger.js";
+import type { Request } from "express";
+
+const storage = multer.diskStorage({
+  destination: (
+    req: Request,
+    file: Express.Multer.File,
+    cb: (error: Error | null, destination: string) => void,
+  ) => {
+    cb(null, "./uploads");
+  },
+  filename: (
+    req: Request,
+    file: Express.Multer.File,
+    cb: (error: Error | null, filename: string) => void,
+  ) => {
+    const fileName = `${Date.now()}-${file.originalname}`;
+    cb(null, fileName);
+  },
+});
+
+const fileFilter = (
+  req: Request,
+  file: Express.Multer.File,
+  cb: multer.FileFilterCallback,
+) => {
+  logger.info(`📁 File mimetype: ${file.mimetype}`);
+  cb(null, true); // Accept everything for now
+};
+
+const limits = {
+  fileSize: 100 * 1024 * 1024,
+};
+
+export const upload = multer({
+  storage: storage,
+  fileFilter: fileFilter,
+  limits: limits,
+});
