+++ package.json
@@ -4,10 +4,11 @@
   "description": "",
   "main": "index.js",
   "scripts": {
-    "api": "node api/server.js",
-    "worker": "node workers/worker.js",
-    "dev:api": "nodemon api/server.js",
-    "dev:worker": "nodemon workers/worker.js"
+    "dev:api": "ts-node-dev --respawn --transpile-only api/server.js",
+    "dev:worker": "ts-node-dev --respawn --transpile-only workers/worker.js",
+    "build": "tsc",
+    "start:api": "node dist/api/server.js",
+    "start:worker": "node dist/workers/worker.js"
   },
   "keywords": [],
   "author": "",
@@ -26,6 +27,12 @@
     "redis": "^5.10.0"
   },
   "devDependencies": {
-    "nodemon": "^3.1.11"
+    "@types/express": "^5.0.6",
+    "@types/multer": "^2.0.0",
+    "@types/node": "^25.3.0",
+    "@types/pino": "^7.0.4",
+    "nodemon": "^3.1.11",
+    "ts-node-dev": "^2.0.0",
+    "typescript": "^5.9.3"
   }
 }
